ai-processor-service/requirements.txt
fastapi==0.95.2
uvicorn==0.20.0
py-eureka-client==0.11.8
python-multipart==0.0.6
pillow==9.5.0
tensorflow==2.10.1
numpy==1.21.6
scikit-learn==1.0.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.13
kafka-python==2.0.2